Discontinue namespacing non-models. Add VERSION, route based on it.
