+++ config/initializers/version.rb
@@ -1,0 +1,7 @@
+# Copyright (c) 2015 The Regents of the University of Michigan.
+# All Rights Reserved.
+# Licensed according to the terms of the Revised BSD License
+# See LICENSE.md for details.
+
+
+VERSION = "1"
