+++ config/routes.rb
@@ -16,7 +16,7 @@ Rails.application.routes.draw do
   mount RailsAdmin::Engine => '/admin', as: 'rails_admin'
 
 
-  namespace :api_v1, path: :"api-v1" do
+  scope "/api-v#{VERSION}" do
     resources :nodes, only: [:index, :show, :create, :update, :destroy], path: :node, param: :namespace
     put "/node/:namespace/auth_credential", controller: :nodes, action: :update_auth_credential
     resources :bags, only: [:index, :show, :create, :update, :destroy], path: :bag, param: :uuid
