+++ spec/helpers/application_helper_spec.rb
@@ -7,8 +7,8 @@ require 'rails_helper'
 
 describe ApplicationHelper do
   before(:each) do
-    allow(ActionController::Metal).to receive(:controller_name).and_return(ApiV1::NodesController.controller_name)
-    allow(ActionController::Metal).to receive(:controller_path).and_return(ApiV1::NodesController.controller_path)
+    allow(ActionController::Metal).to receive(:controller_name).and_return(NodesController.controller_name)
+    allow(ActionController::Metal).to receive(:controller_path).and_return(NodesController.controller_path)
   end
 
   describe "assignee" do
@@ -34,7 +34,7 @@ describe ApplicationHelper do
 
   describe "adapter" do
     it "finds the correct adapter" do
-      expect(adapter).to eql(ApiV1::NodeAdapter)
+      expect(adapter).to eql(NodeAdapter)
     end
   end
